+++ datacraft/server.py
@@ -1,6 +1,7 @@
 """
 Light weight module for running a Flask Server that returns data from a generator.
 """
+import time
 from typing import Generator
 import logging
 import flask
@@ -19,11 +20,13 @@ class _Server:
                  generator: Generator,
                  endpoint: str,
                  data_is_json: bool,
-                 count_supplier: datacraft.ValueSupplierInterface):
+                 count_supplier: datacraft.ValueSupplierInterface,
+                 delay: float = None):
         self.generator = generator
         self.endpoint = endpoint
         self.data_is_json = data_is_json
         self.count_supplier = count_supplier
+        self.delay = delay
         self.call_number = 0
 
     def callback(self):
@@ -40,6 +43,8 @@ class _Server:
         except StopIteration:
             _log.warning('No more iterations available')
             return flask.Response(None, status=204)
+        if self.delay:
+            time.sleep(self.delay)
         if self.data_is_json:
             return flask.jsonify(data)
         # this may already be json or templated data
@@ -59,7 +64,8 @@ class _Server:
 def run(generator: Generator,
         endpoint: str,
         data_is_json: bool,
-        count_supplier: datacraft.ValueSupplierInterface):
+        count_supplier: datacraft.ValueSupplierInterface,
+        delay: float = None):
     """
     Runs a light weight Flask server with data returned by the provided generator served at each subsequent call to
     the provided endpoint. End point should start with /. When StopIteration encountered, returns a 204 status code.
@@ -69,6 +75,7 @@ def run(generator: Generator,
         endpoint: to serve data on i.e. /data
         data_is_json: if the data should be returned as JSON, default is as string
         count_supplier: supplies the number of values to generate for each call
+        delay: number of seconds to pause between request and response
     """
-    server = _Server(generator, endpoint, data_is_json, count_supplier)
+    server = _Server(generator, endpoint, data_is_json, count_supplier, delay)
     server.run()